opt: gpu buffer more abstract

+++ Src/Vocore.Graphics/Descriptor/BufferDescriptor.cs
@@ -11,7 +11,7 @@ public struct BufferDescriptor
     /// <param name="size">The size of the buffer.</param>
     /// <param name="usage">The usage flags for the buffer.</param>
     /// <param name="name">The name of the buffer (optional).</param>
-    public BufferDescriptor(ulong size, BufferUsage usage, string name = "unnamed_gpu_buffer")
+    public BufferDescriptor(uint size, BufferUsage usage, string name = "unnamed_gpu_buffer")
     {
         Size = size;
         Usage = usage;
@@ -21,7 +21,7 @@ public struct BufferDescriptor
     /// <summary>
     /// The size of the buffer.
     /// </summary>
-    public ulong Size { get; init; }
+    public uint Size { get; init; }
 
 
     /// <summary>
